* encoding detection: format-fixes (avoid assertion with wx2.9)
git-svn-id: https://svn.code.sf.net/p/codeblocks/code/trunk@8892 2a5c6006-c6dd-42ca-98ab-0921f2732cef

+++ src/sdk/encodingdetector.cpp
@@ -489,7 +489,7 @@ bool EncodingDetector::ConvertToWxString(const wxByte* buffer, size_t size)
             if (m_UseLog && outlen>0)
             {
                 logmsg.Printf(_T("Conversion succeeded using wxEncodingConverter "
-                                 "(buffer size = %d, converted size = %d."), size, outlen);
+                                 "(buffer size = %lu, converted size = %lu."), static_cast<unsigned long>(size), static_cast<unsigned long>(outlen));
                 logmgr->DebugLog(logmsg);
             }
         }
@@ -503,7 +503,7 @@ bool EncodingDetector::ConvertToWxString(const wxByte* buffer, size_t size)
                 if (m_UseLog && outlen>0)
                 {
                     logmsg.Printf(_T("Conversion succeeded using wxCSConv "
-                                     "(buffer size = %d, converted size = %d."), size, outlen);
+                                     "(buffer size = %lu, converted size = %lu."), static_cast<unsigned long>(size), static_cast<unsigned long>(outlen));
                     logmgr->DebugLog(logmsg);
                 }
             }
